+++ packages/svgkit/src/document/serialize.js
