+++ packages/svgkit/src/page/element.js
