+++ packages/svgkit/src/utils/getPageSize.js
